Add files via upload
I have successfully developed an advanced dashboard that serves as a powerful tool for dissecting and interpreting startup funding data. This dynamic platform offers a sophisticated analysis of startup trends on various dimensions, including country, city, and continent-specific insights. Through meticulous examination, I've provided a detailed breakdown of the average number of years it takes for startups to achieve the coveted billion-dollar status from their founding date.

In addition to geographical and temporal analyses, I've delved into the intricacies of industry-wise trends. By categorizing startup data based on both geographical locations (city and continent) and age, the dashboard offers a nuanced perspective on which industries thrive in specific countries. This multi-faceted approach allows for a deeper understanding of the global startup ecosystem.

To further enhance the dashboard's value, I've incorporated a comprehensive assessment of the average valuation of startups up to the year 2022. This strategic addition not only enriches our understanding of the startup landscape but also provides valuable insights for stakeholders and decision-makers in the startup ecosystem.
 main